% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_trim.R
\name{get_cl_to_trim}
\alias{get_cl_to_trim}
\alias{suggest_trim}
\title{Cluster-wise trimming diagnostics and suggested trim levels}
\usage{
get_cl_to_trim(data, vec_var, formula)

suggest_trim(data, vec_var = "Vj", cl_var = "cluster", formula = "x + I(x^2)")
}
\arguments{
\item{data}{A data frame.}

\item{vec_var}{String. Name of the numeric column to trim (e.g., \code{"Vj"}).}

\item{formula}{String. RHS of a formula in terms of \code{x}
(e.g., \code{"x + I(x^2)"}). The response is internally set to \code{val}.}

\item{cl_var}{String. Cluster column name (default \code{"cluster"}).}
}
\value{
\itemize{
\item \code{get_cl_to_trim()}: A numeric vector of average absolute errors
(same length/order as \code{j_grid}).
\item \code{suggest_trim()}: A list with:
\describe{
\item{\code{original_data}}{Input \code{data} with an added/standardized \code{cluster} column.}
\item{\code{data_for_trimming}}{Tibble with columns \code{cluster}, \code{trim_level}, \code{abs_error}.}
}
}
}
\description{
A funciton that computes trimmed means
}
\examples{
# Example:
# suggest_trim(dat, vec_var = "Vj", cl_var = "cluster", formula_rhs = "x + I(x^2)")

}
